simclr-rhd/requirements.txt

google-cloud-storage
matplotlib
numpy
pandas
pillow
scikit-learn
scipy
torch
torchvision
umap-learn